myems-mqtt/requirements.txt
mysql-connector-python==9.6.0
paho-mqtt==2.1.0
schedule
python-decouple==3.8